fix(android): keep panning the caret past the readable window
Editors that expose no ExtractedText left the pan dead once the caret sat
more than a thousand characters into the field: the readable window is the
only thing the resolver could measure, and a full one says nothing about
where the caret actually is. So it refused, correctly but uselessly.

The host reports the caret through onUpdateSelection, which is an absolute
position and the reference that was missing. It survives the start of a pan
because finishComposingText does not move the caret, and the resolver's own
setSelection keeps it current for every step after the first. A report from
behind the text we can see is stale beyond rescuing and still refuses.
